net-cli: drop the redundant `as u8` on CliError::code().
`CliError::code()` already returns `u8`; the cast to `u8` was a no-op that
tripped clippy::unnecessary_cast.

+++ net/crates/net/cli/src/main.rs
@@ -180,7 +180,7 @@ async fn main() -> ExitCode {
             // Print the kind + message to stderr; the exit code
             // carries the discriminator for scripting consumers.
             eprintln!("net: {}", e);
-            ExitCode::from(e.code() as u8)
+            ExitCode::from(e.code())
         }
     }
 }
